Move ScenarioComponentsUsages to RestModel
