+++ designer/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/processdetails.scala
@@ -10,7 +10,7 @@ import pl.touk.nussknacker.engine.api.process.{ProcessName, VersionId, ProcessId
 import pl.touk.nussknacker.engine.api.{ProcessVersion => EngineProcessVersion}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.restmodel.displayedgraph.{DisplayableProcess, ValidatedDisplayableProcess}
-import pl.touk.nussknacker.restmodel.process.{ProcessIdWithName, ProcessingType}
+import pl.touk.nussknacker.restmodel.process.{ProcessIdWithName, ProcessingType, ScenarioComponentsUsages}
 
 import java.time.Instant
 
@@ -119,6 +119,7 @@ object processdetails {
     implicit case object FetchCanonical extends ProcessShapeFetchStrategy[CanonicalProcess]
     // In fact Unit won't be returned inside shape and Nothing would be more verbose but it won't help in compilation because Nothing <: DisplayableProcess
     implicit case object NotFetch extends ProcessShapeFetchStrategy[Unit]
+    implicit case object FetchComponentsUsages extends ProcessShapeFetchStrategy[ScenarioComponentsUsages]
   }
 
   type ProcessDetails = BaseProcessDetails[DisplayableProcess]
