+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/db/entity/ProcessVersionEntityFactory.scala
@@ -9,7 +9,7 @@ import pl.touk.nussknacker.engine.api.process.{ProcessId, VersionId}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.marshall.ProcessMarshaller
 import pl.touk.nussknacker.restmodel.component.NodeId
-import pl.touk.nussknacker.ui.process.repository.{ComponentIdParts, ScenarioComponentsUsages}
+import pl.touk.nussknacker.restmodel.process.{ComponentIdParts, ScenarioComponentsUsages}
 import slick.lifted.{ForeignKeyQuery, ProvenShape, TableQuery => LTableQuery}
 import slick.sql.SqlProfile.ColumnOption.NotNull
